backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
MetaTrader5>=5.0.45
pydantic==2.6.1
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
sqlalchemy==2.0.27
python-telegram-bot>=20.8
numpy>=1.26.4
websockets>=12.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.5

pandas>=2.2.0
